+++ src/utils.ts
@@ -21,13 +21,13 @@ export const generateShareUrl = async (origin: string, openapi?: string, config?
 
     if (config.sortSet !== undefined) configOps.sortSet = await stringify(config.sortSet);
     if (config.filterSet !== undefined) configOps.filterSet = await stringify(config.filterSet);
-    // if (options.casingSet !== undefined) config.casingSet = await stringify(options.casingSet);
+    if (config.generateSet !== undefined) configOps.generateSet = await stringify(config.generateSet);
+    if (config.casingSet !== undefined) configOps.casingSet = await stringify(config.casingSet);
     if (config.sort !== undefined) configOps.sort = config.sort;
     // if (options.rename !== undefined) config.rename = options.rename;
     // if (options.convertTo !== undefined) config.convertTo = options.convertTo
 
     if (config.isFilterOptionsCollapsed !== undefined) configOps.isFilterOptionsCollapsed = config.isFilterOptionsCollapsed;
-    if (config.isSortOptionsCollapsed !== undefined) configOps.isSortOptionsCollapsed = config.isSortOptionsCollapsed;
     if (config.outputLanguage !== undefined) configOps.outputLanguage = config.outputLanguage;
 
     if (config.pathSort !== undefined) configOps.pathSort = config.pathSort;